requirements.txt
pyyaml==6.0.2
pandas==2.2.3
matplotlib==3.9.2
